Match ParseInt return type.

+++ lock_unix.go
@@ -9,7 +9,7 @@ import (
 
 // uidGidInt returns uid, gid of the user as a number.
 func uidGidInt(u user.User) (uid, gid uint32, err error) {
-	var ui, gi int
+	var ui, gi int64
 	ui, err = strconv.ParseInt(u.Uid, 10, 32)
 	if err != nil {
 		return
